backend/requirements.txt
fastapi
uvicorn
numpy
faster-whisper
sentence-transformers
librosa
groq
python-dotenv
deepface
opencv-python
mediapipe
tf-keras
motor
pypdf
python-multipart
reportlab
deepface
opencv-python
pillow
librosa
soundfile
scipy
keybert
spacy
